Check and update ligthning_ndk based on stored release tag name
